requirements.txt
fastapi
granian
python-dotenv
web3==6.*
redis
aiohttp
tronpy
asgiref
tonutils